requirements.txt
# InterActVLM-Discrete (IVD) Dependencies
# 使用镜像: uv pip install -r requirements.txt -i https://pypi.tuna.tsinghua.edu.cn/simple/

# Core
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scipy>=1.10.0

# Vision Transformer & VLM
transformers>=4.36.0
peft>=0.7.0  # LoRA support
accelerate>=0.25.0

# 3D Rendering
# pytorch3d 需要从源码编译，见 README
trimesh>=4.0.0

# Image processing
Pillow>=10.0.0
opencv-python>=4.8.0
albumentations>=1.3.0

# Configuration
omegaconf>=2.3.0

# Logging & Monitoring
tensorboard>=2.14.0
tqdm>=4.65.0

# Utilities
einops>=0.7.0
timm>=0.9.0

# Metrics
scikit-learn>=1.3.0

# YAML
pyyaml>=6.0
